backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shrpow", email = "shrpow@ya.ru" }
]
dependencies = [
    "beanie>=1.26.0",
    "jinja2>=3.1.4",
    "apscheduler>=3.10.4",
    "loguru>=0.7.2",
    "aiogram>=3.13.1",
    "prometheus-client>=0.21.0",
    "pyzmq>=26.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]
